# documentation/openapi.yaml
openapi: 3.1.0
info:
  title: EasyList API
  description: |
    JSON:API совместимый REST API для управления вакансиями, контактами, документами и комментариями.

    Все ответы используют формат JSON:API (`application/vnd.api+json`).
    Авторизация через Bearer-токен (Laravel Sanctum).
  version: 1.0.0
  contact:
    name: EasyList Team
servers:
  - url: /api/v1
    description: API v1
...
security:
  - bearerAuth: []
paths:
...
  /jobs/{jobId}/favorite:
    patch:
      operationId: toggleFavorite
      summary: Переключить статус избранного
      tags: [Jobs]
      parameters:
        - $ref: '#/components/parameters/JobId'
      responses:
        '200':
          description: Статус избранного переключён
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/SingleJobResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Laravel Sanctum Bearer Token
  parameters:
    JobId:
      name: jobId
      in: path
      required: true
      schema:
        type: integer
...
  responses:
    Unauthorized:
      description: Не авторизован
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: Unauthenticated.
    Forbidden:
      description: Доступ запрещён
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
...
  schemas:
...
    JobResource:
      type: object
      properties:
        type:
          type: string
          example: jobs
        id:
          type: string
          example: '42'
        attributes:
          type: object
          properties:
            uuid:
              type: string
              format: uuid
              nullable: true
            title:
              type: string
            company_name:
              type: string
            description:
              type: string
              nullable: true
            job_url:
              type: string
              format: uri
              nullable: true
            location_city:
              type: string
              nullable: true
            salary:
              type: integer
              nullable: true
            is_favorite:
              type: boolean
            created_at:
              type: string
              format: date-time
        relationships:
          type: object
          properties:
            status:
              $ref: '#/components/schemas/RelationshipLinkage'
            category:
              $ref: '#/components/schemas/RelationshipLinkage'
            skills:
              type: object
              properties:
                data:
                  type: array
                  items:
                    $ref: '#/components/schemas/ResourceIdentifier'
            contacts:
              type: object
              properties:
                data:
                  type: array
                  items:
                    $ref: '#/components/schemas/ResourceIdentifier'
            comments:
              type: object
              properties:
                data:
                  type: array
                  items:
                    $ref: '#/components/schemas/ResourceIdentifier'
            documents:
              type: object
              properties:
                data:
                  type: array
                  items:
                    $ref: '#/components/schemas/ResourceIdentifier'
            tasks:
              type: object
              properties:
                data:
                  type: array
                  items:
                    $ref: '#/components/schemas/ResourceIdentifier'
    JobStatusResource:
      type: object
      properties:
        type:
          type: string
          example: job-statuses
        id:
          type: string
        attributes:
          type: object
          properties:
            title:
              type: string
            description:
              type: string
              nullable: true
            color:
              type: string
              description: Цвет статуса (enum)
            order_column:
              type: integer
    JobCategoryResource:
      type: object
      properties:
        type:
          type: string
          example: job-categories
        id:
          type: string
        attributes:
          type: object
          properties:
            title:
              type: string
            description:
              type: string
              nullable: true
            currency:
              type: string
              description: Код валюты (enum)
            currency_symbol:
              type: string
            order_column:
              type: integer
    CommentResource:
      type: object
      properties:
        type:
          type: string
          example: comments
        id:
          type: string
        attributes:
          type: object
          properties:
            body:
              type: string
            author_name:
              type: string
            user_id:
              type: integer
            created_at:
              type: string
              format: date-time
    DocumentResource:
      type: object
      properties:
        type:
          type: string
          example: documents
        id:
          type: string
        attributes:
          type: object
          properties:
            title:
              type: string
            description:
              type: string
              nullable: true
            category:
              type: string
              description: Категория документа (resume, cover_letter, portfolio, etc.)
            category_label:
              type: string
            file_path:
              type: string
              nullable: true
            original_filename:
              type: string
              nullable: true
            mime_type:
              type: string
              nullable: true
            file_size:
              type: integer
              nullable: true
            external_url:
              type: string
              format: uri
              nullable: true
            author_name:
              type: string
            created_at:
              type: string
              format: date-time
    ContactResource:
      type: object
      properties:
        type:
          type: string
          example: contacts
        id:
          type: string
        attributes:
          type: object
          properties:
            first_name:
              type: string
            last_name:
              type: string
            position:
              type: string
              nullable: true
            city:
              type: string
              nullable: true
            email:
              type: string
              format: email
              nullable: true
            phone:
              type: string
              nullable: true
            description:
              type: string
              nullable: true
            linkedin_url:
              type: string
              format: uri
              nullable: true
            facebook_url:
              type: string
              format: uri
              nullable: true
            whatsapp_url:
              type: string
              format: uri
              nullable: true
            created_at:
              type: string
              format: date-time
    TaskResource:
      type: object
      properties:
        type:
          type: string
          example: tasks
        id:
          type: string
        attributes:
          type: object
          properties:
            title:
              type: string
            user_id:
              type: integer
            external_id:
              type: string
              nullable: true
            deadline:
              type: string
              format: date-time
              nullable: true
            completed_at:
              type: string
              format: date-time
              nullable: true
            created_at:
              type: string
              format: date-time
    CompanyInfoResource:
      type: object
      properties:
        type:
          type: string
          example: company-info
        id:
          type: string
        attributes:
          type: object
          properties:
            name:
              type: string
            city:
              type: string
              nullable: true
            info:
              type: object
              nullable: true
              description: Детальная информация о компании от ИИ
    SkillResource:
      type: object
      properties:
        type:
          type: string
          example: skills
        id:
          type: string
        attributes:
          type: object
          properties:
            title:
              type: string
...
    ResourceIdentifier:
      type: object
      properties:
        type:
          type: string
        id:
          type: string
    RelationshipLinkage:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/ResourceIdentifier'

    # --- Pagination ---
...
    SingleJobResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/JobResource'
        included:
          type: array
          items:
            oneOf:
              - $ref: '#/components/schemas/JobStatusResource'
              - $ref: '#/components/schemas/JobCategoryResource'
              - $ref: '#/components/schemas/SkillResource'
              - $ref: '#/components/schemas/ContactResource'
              - $ref: '#/components/schemas/CommentResource'
              - $ref: '#/components/schemas/DocumentResource'
              - $ref: '#/components/schemas/TaskResource'
              - $ref: '#/components/schemas/CompanyInfoResource'

    # --- Request Bodies ---
...